requirements.txt
python-telegram-bot
python-dotenv
pytest-asyncio
pytest
telethon
beautifulsoup4
markdown
mysql-connector-python
pymongo==3.6
pymongo[srv]
opencv-python-headless
locust